requirements.txt
certifi==2022.12.7
cfbd==4.4.12
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tabulate==0.9.0
urllib3==1.26.15